# docs/pkg/pb/dipan_type.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "pkg/pb/dipan_type.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "DipanTypeService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/dipan-types": {
      "get": {
        "operationId": "DipanTypeService_ListDipanTypes",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbListDipanTypesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "DipanTypeService"
        ]
      },
      "post": {
        "operationId": "DipanTypeService_CreateDipanType",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbDipanType"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbCreateDipanTypeRequest"
            }
          }
        ],
        "tags": [
          "DipanTypeService"
        ]
      }
    },
    "/v1/dipan-types/{id}": {
      "get": {
        "operationId": "DipanTypeService_GetDipanType",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbDipanType"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "DipanTypeService"
        ]
      },
      "delete": {
        "operationId": "DipanTypeService_DeleteDipanType",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbDeleteDipanTypeResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "DipanTypeService"
        ]
      },
      "put": {
        "operationId": "DipanTypeService_UpdateDipanType",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbDipanType"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DipanTypeServiceUpdateDipanTypeBody"
            }
          }
        ],
        "tags": [
          "DipanTypeService"
        ]
      }
    }
  },
  "definitions": {
    "DipanTypeServiceUpdateDipanTypeBody": {
      "type": "object",
      "properties": {
        "namaType": {
          "type": "string"
        }
      }
    },
    "pbCreateDipanTypeRequest": {
      "type": "object",
      "properties": {
        "namaType": {
          "type": "string"
        }
      }
    },
    "pbDeleteDipanTypeResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      }
    },
    "pbDipanType": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "namaType": {
          "type": "string"
        }
      }
    },
    "pbListDipanTypesResponse": {
      "type": "object",
      "properties": {
        "dipanTypes": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/pbDipanType"
          }
        },
        "total": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}